Fixed jsch broken pipe exception
I was executing 2 commands sequentially and the second command needed sudo access. Appending sudo to the second command fixed this part. Then again I was downloading binaries for maven 3.3.3 and trying to move a folder named apache-maven-3.2.3 which was causing the same error at this step. And that's why you never do a copy paste git status sigh